requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
Pillow>=7.1.2
djangorestframework>=3.11.0
django-filter>=0.11.0